Removed duplicate function 5. Added spacing to functoin 5 docstring
